requirements.txt
alembic==1.18.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
google-auth==2.48.0
google-auth-oauthlib==1.2.4
greenlet==3.3.1
gspread==6.2.1
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.19
typing_extensions==4.15.0
urllib3==2.6.3
waitress==2.1.2
Werkzeug==3.1.5
WTForms==3.2.1
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
